fix Raire simulations for estimation
