+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/audit/TestRaireAudit.kt
@@ -19,7 +19,7 @@ class TestRaireAudit {
     }
 
     fun testRaireWorkflow(auditConfig: AuditConfig) {
-        val (rcontest, testCvrs) = simulateRaireTestData(N=20000, contestId=111, ncands=3, minMargin=.04, quiet = true)
+        val (rcontest, testCvrs) = simulateRaireTestContest(N=20000, contestId=111, ncands=3, minMargin=.04, quiet = true)
         val workflow = ClcaAudit(auditConfig, emptyList(), listOf(rcontest),
             MvrManagerClcaForTesting(testCvrs, testCvrs, auditConfig.seed))
         runAudit("testRaireWorkflow", workflow)
@@ -33,7 +33,7 @@ class TestRaireAudit {
             clcaConfig = ClcaConfig(ClcaStrategyType.fuzzPct, simFuzzPct = mvrFuzzPct)
         )
 
-        val (rcontest: RaireContestUnderAudit, testCvrs: List<Cvr>) = simulateRaireTestData(N=20000, contestId=111, ncands=4, minMargin=.04, quiet = true)
+        val (rcontest: RaireContestUnderAudit, testCvrs: List<Cvr>) = simulateRaireTestContest(N=20000, contestId=111, ncands=4, minMargin=.04, quiet = true)
         val testMvrs = makeFuzzedCvrsFrom(listOf(rcontest.contest), testCvrs, mvrFuzzPct)
         val workflow = ClcaAudit(auditConfig, emptyList(), listOf(rcontest),
             MvrManagerClcaForTesting(testCvrs, testMvrs, auditConfig.seed))
